+++ src/components/NavigatorSearch/modes/default/views.ts
@@ -9,7 +9,7 @@ import { sectionNames } from "@dashboard/intl";
 import { fuzzySearch } from "@dashboard/misc";
 import { menuListUrl } from "@dashboard/navigation/urls";
 import { orderDraftListUrl, orderListUrl } from "@dashboard/orders/urls";
-import { pageListUrl } from "@dashboard/pages/urls";
+import { pageListUrl } from "@dashboard/modeling/urls";
 import { permissionGroupListUrl } from "@dashboard/permissionGroups/urls";
 import { pluginListUrl } from "@dashboard/plugins/urls";
 import { productListUrl } from "@dashboard/products/urls";
