+++ src/index.tsx
@@ -65,8 +65,8 @@ import NavigationSection from "./navigation";
 import { structuresListPath } from "./navigation/urls";
 import { NotFound } from "./NotFound";
 import OrdersSection from "./orders";
-import PageSection from "./pages";
-import { pageListPath } from "./pages/urls";
+import PageSection from "./modeling";
+import { pageListPath } from "@dashboard/modeling/urls";
 import PageTypesSection from "./pageTypes";
 import { pageTypeListPath } from "./pageTypes/urls";
 import PermissionGroupSection from "./permissionGroups";
