+++ src/modeling/components/PageDetailsPage/PageDetailsPage.tsx
@@ -26,7 +26,7 @@ import { useBackLinkWithState } from "@dashboard/hooks/useBackLinkWithState";
 import useDateLocalize from "@dashboard/hooks/useDateLocalize";
 import { SubmitPromise } from "@dashboard/hooks/useForm";
 import useNavigator from "@dashboard/hooks/useNavigator";
-import { modelingSection } from "@dashboard/pages/urls";
+import { modelingSection } from "@dashboard/modeling/urls";
 import { FetchMoreProps, RelayToFlat } from "@dashboard/types";
 import { mapNodeToChoice } from "@dashboard/utils/maps";
 import React from "react";
