+++ src/modeling/components/PageListDatagrid/PageListDatagrid.tsx
@@ -6,8 +6,8 @@ import {
   useDatagridChangeState,
 } from "@dashboard/components/Datagrid/hooks/useDatagridChange";
 import { TablePaginationWithContext } from "@dashboard/components/TablePagination";
-import { Page, Pages } from "@dashboard/pages/types";
-import { PageListUrlSortField } from "@dashboard/pages/urls";
+import { Page, Pages } from "@dashboard/modeling/types";
+import { PageListUrlSortField } from "@dashboard/modeling/urls";
 import { ListProps, SortPage } from "@dashboard/types";
 import { Item } from "@glideapps/glide-data-grid";
 import { Box, useTheme } from "@saleor/macaw-ui-next";
