+++ src/modeling/utils/data.ts
