+++ src/modeling/utils/validation.ts
