+++ src/modeling/views/PageDetails.tsx
